refactor: move `PackageOrRockspec` to `args` module
